Big commit, had to reset commits with api key and client-id.
Changed around in gamecontroller so it only gets games from api. Created IGDBService for fetching games by title from IGDB API with limit and offset params.
Made it so api key and client id are set to system variables. Also did this in github workflow and added secrets to github.